Add Axios to project
Added Axios to package.json and created a custom axios instance in `lib/axiosConfig.js` that exports the instance with some defaults.

This serves as a workaround for the issue of `globalProperties` not behaving as expected in the new Vue3 ecosystem.

See this SO question for more detail of the issue at hand:
https://stackoverflow.com/questions/66711937/vuex-4-modules-cant-use-global-axios-property